+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/state/extension/WebExtensionPromptRequest.kt
@@ -4,6 +4,7 @@
 
 package mozilla.components.browser.state.state.extension
 
+import mozilla.components.concept.engine.webextension.PermissionPromptResponse
 import mozilla.components.concept.engine.webextension.WebExtension
 import mozilla.components.concept.engine.webextension.WebExtensionInstallException
 
@@ -43,19 +44,19 @@ sealed class WebExtensionPromptRequest {
          */
         sealed class Permissions(
             override val extension: WebExtension,
-            open val onConfirm: (Boolean) -> Unit,
         ) : AfterInstallation(extension) {
             /**
              * Value type that represents a request for a required permissions prompt.
              * @property extension The [WebExtension] that requested the dialog to be shown.
              * @property permissions The permissions to list in the dialog.
-             * @property onConfirm A callback indicating whether the permissions were granted or not.
+             * @property onConfirm A callback indicating the prompt has been confirmed and pass
+             * [PermissionPromptResponse] result.
              */
             data class Required(
                 override val extension: WebExtension,
                 val permissions: List<String>,
-                override val onConfirm: (Boolean) -> Unit,
-            ) : Permissions(extension, onConfirm)
+                val onConfirm: (PermissionPromptResponse) -> Unit,
+            ) : Permissions(extension)
 
             /**
              * Value type that represents a request for an optional permissions prompt.
@@ -66,8 +67,8 @@ sealed class WebExtensionPromptRequest {
             data class Optional(
                 override val extension: WebExtension,
                 val permissions: List<String>,
-                override val onConfirm: (Boolean) -> Unit,
-            ) : Permissions(extension, onConfirm)
+                val onConfirm: (Boolean) -> Unit,
+            ) : Permissions(extension)
         }
 
         /**